Reformatting files after convo with Ozer
Reorganizing files

renaming files

replacing names

renaming files for base

Updating resource file

Updating package names

Updating naming of directories

Renaming and beginning refactor

Moving types
