+++ behaviours/go_to_geopoint/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/go_to_geopoint
+[install]
+install_scripts=$base/lib/go_to_geopoint
