--- behaviours/go_to_setpoint/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/go_to_setpoint
-[install]
-install_scripts=$base/lib/go_to_setpoint
