+++ behaviours/smarc_action_base/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>smarc_action_client</name>
+  <name>smarc_action_base</name>
   <version>0.0.0</version>
   <description>SMARC simplified action-client interface</description>
   <maintainer email="kogucki@kth.se">tko</maintainer>
