+++ behaviours/smarc_action_base/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/smarc_action_base
+[install]
+install_scripts=$base/lib/smarc_action_base
