+++ behaviours/smarc_action_base/smarc_action_base/smarc_action_base.py
@@ -1,6 +1,4 @@
 import abc
-from argparse import Action
-from typing import TypeVar
 
 # ROS Imports
 from rclpy.action import ActionClient, ActionServer, GoalResponse, CancelResponse
@@ -11,7 +9,60 @@ from rclpy.action.server import ServerGoalHandle
 from rclpy.node import Node
 from rclpy.type_support import check_for_type_support
 
-from smarc_action_client.smarc_ros_types import ActionType, ActionFeedback, ActionGoal, ActionResult
+from smarc_action_base.smarc_ros_types import ActionFeedback, ActionGoal, ActionResult
+
+class ActionType:
+    """Wrapper around ROS Action Type to provide easy dot completion.
+
+    Attributes:
+        Result: empty result message
+        Feedback: empty feedback message
+        Goal: empty goal message
+    """
+
+    def __init__(self, action_type):
+        self._action_type = action_type
+        self.validate_type()
+
+    def validate_type(self):
+        """Evaluates whether provided action type is a valid ROS action type.
+
+        Raises:
+            AttributeError: Provides additional context to user to help debug ROS error.
+        """
+        try:
+            check_for_type_support(self._action_type)
+        except AttributeError as err:
+            err_str = "Provided action_type is not a valid ROS action type.\n"
+            err_str += "Action types generally should be of type `from some_interface.action import MyAction"
+            raise AttributeError(err_str) from err
+
+
+    @property
+    def ros_type(self):
+        """Underlying ROS type.
+
+        Returns:
+            action: ROS action type
+        """
+        return self._action_type
+
+    @property
+    def Result(self) -> ActionResult:
+        """Empty results message."""
+        return self._action_type.Result()
+
+    @property
+    def Feedback(self) -> ActionFeedback:
+        """Empty feedback message."""
+        return self._action_type.Feedback()
+
+    @property
+    def Goal(self) -> ActionGoal:
+        """Empty goal message."""
+        return self._action_type.Goal()
+
+
 
 class SMARCActionServer(abc.ABC):
     """Action Server base class
