+++ behaviours/smarc_action_base/smarc_action_base/smarc_ros_types.py
@@ -1,0 +1,25 @@
+from typing import Protocol
+
+
+class Msg(Protocol):
+    """Generic Message Alias."""
+
+    pass
+
+
+class ActionResult(Msg, Protocol):
+    """Action Type Result Message"""
+
+    pass
+
+
+class ActionFeedback(Msg, Protocol):
+    """Action Type Feedback Message"""
+
+    pass
+
+
+class ActionGoal(Msg, Protocol):
+    """Action Type Goal Message"""
+
+    pass
