--- utilities/smarc_action_client/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/smarc_action_client
-[install]
-install_scripts=$base/lib/smarc_action_client
